requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Script==0.5.3
Jinja2==2.7
MarkupSafe==0.18
Twisted==13.1.0
Werkzeug==0.9.1
amqp==1.0.12
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
celery==3.0.21
celery-with-redis==3.0
itsdangerous==0.22
kombu==2.5.12
pyjade==2.0.2
pyserial==2.6
python-dateutil==2.1
redis==2.7.6
six==1.3.0
uWSGI==1.9.14
webassets==0.9.dev
wsgiref==0.1.2
zope.interface==4.0.5
